app_code/requirements.txt
streamlit==1.38.0
requests
databricks-sdk
pandas